refactor(game): organize terrain system files
